pyproject.toml
[project]
name = "telegentic"
version = "0.1.0"
description = "Minimal and elegant Telegram bot framework"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["aiogram>=3.20.0", "python-dotenv>=1.0.0", "pydantic>=2.0.0"]

[project.optional-dependencies]
fastapi = ["fastapi>=0.115.0", "uvicorn>=0.30.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "black>=25.1.0",
    "ipykernel>=6.29.5",
    "isort>=6.0.1",
    "mypy>=1.16.1",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=6.0.0",
    "telegentic[fastapi]",
]

[tool.uv.sources]
telegentic = { workspace = true }

[tool.mypy]
python_version = "3.11"
packages = ["telegentic", "tests"]
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
